added email verification
user can verify their email anytime but has limited access

+++ backend/src/services/book.service.ts
@@ -6,6 +6,9 @@ export async function createBook(data: BookCreationDto){
 
     const newBook = new Book({...data});
     const savedBook = await newBook.save();
+    if( !savedBook ){
+        throw new Error("Book not saved");
+    }
     return savedBook;   
 }
 
@@ -23,22 +26,62 @@ export async function updateBook(bookId:string, data: BookUpdateDto){
 
 export async function deleteBook(id: string) {
     const deletedChapter = await Chapter.deleteMany({ bookId: id });
-    const deletedBook = await Book.findByIdAndDelete( id );
-    return deletedBook;
+    const result = await Book.findByIdAndDelete( id );
+    if (!result) {
+        throw new Error("Book not found");
+    }
+    return result;
 }
 
 //may delete all chapters not may be based on query
 export async function deleteAllBooks(query: any = {} ) {
+
     const deletedChapter = await Chapter.deleteMany();
-    const deletedBook =await Book.deleteMany(query);
-    return deletedBook;
+    if ( deletedChapter.deletedCount === 0) {
+        throw new Error("No Chapters matched the criteria for deletion.");
+    }
+    const result =await Book.deleteMany(query);
+    if ( result.deletedCount === 0) {
+        throw new Error("No books matched the criteria for deletion.");
+    }
+    return result;
+}
+
+export async function findBook(id: string) {
+    const result = await Book.findById(id);
+    if (!result) {
+        throw new Error("Book not found");
+    }
+    return result;
 }
 
 export async function countBooks(query:any = {}) {
   return await Book.countDocuments(query);
 }
 
-export async function findBook(id: string) {
-   const result = await Book.findOne( {_id: id} );
-    return result;
+
+//returns 
+//books-> name id image  
+//and its three latest chapters
+//  with their name chapterNo createdAt
+export async function findBooks(pageNo: number) {
+
+    const limit = 20;
+    const skip = ( pageNo - 1 ) * limit; //e.g. for 3rd page // 3-1 * 10 => 20 books
+
+    const books = await Book.find()
+        .sort({order: -1})
+        .skip(skip)
+        .limit(limit)
+        .lean();    //this helps when adding custom fields manually
+    
+    for (const book of books) {
+        const chapters = await Chapter.find({ bookId: book._id})
+            .sort({order: -1})  //same order as books
+            .limit(3)           //gives 3 chapters
+            .select("name chapterNumber createdAt")
+            .lean();    
+        (book as any).latestChapters = chapters; //treats as plain js object and adds latestChapters to it 
+    };  
+    return books;
 }
